requirements.txt
python-dotenv
google-generativeai
langchain 
faiss--cpu
streamlit
-e .